src: cast strlen() to uint16_t

+++ src/btstack_util.c
@@ -630,7 +630,7 @@ uint16_t btstack_next_cid_ignoring_zero(uint16_t current_cid){
 }
 
 uint16_t btstack_strcpy(char * dst, uint16_t dst_size, const char * src){
-    uint16_t bytes_to_copy = (uint16_t) btstack_min( dst_size - 1, strlen(src));
+    uint16_t bytes_to_copy = (uint16_t) btstack_min( dst_size - 1, (uint16_t) strlen(src));
     (void) memcpy(dst, src, bytes_to_copy);
     dst[bytes_to_copy] = 0;
     return bytes_to_copy + 1;
